+++ CHANGELOG.md
@@ -1,5 +1,50 @@
 # Changelog
 
+## [0.81.1](https://github.com/zarf-dev/zarf/compare/v0.81.0...v0.81.1) (2026-07-17)
+
+
+### Bug Fixes
+
+* change adoptExistingResources to takeOwnership ([#5010](https://github.com/zarf-dev/zarf/issues/5010)) ([c5eb289](https://github.com/zarf-dev/zarf/commit/c5eb289bcd535979d47e1b8311744d5018929479))
+* **helm:** create registry client for http chart with OCI re-direct ([#5067](https://github.com/zarf-dev/zarf/issues/5067)) ([34d0cf1](https://github.com/zarf-dev/zarf/commit/34d0cf1a8322a70a8e1f3605038d64dfeb9401e4))
+* **mTLS:** prevent negotiation when portforward with mTLS ([#5054](https://github.com/zarf-dev/zarf/issues/5054)) ([bbce3c3](https://github.com/zarf-dev/zarf/commit/bbce3c3a3da5157346b4afac0655cebf1ba8ebd7))
+* **proxy:** support for system CA with mTLS certs ([#5074](https://github.com/zarf-dev/zarf/issues/5074)) ([93f4719](https://github.com/zarf-dev/zarf/commit/93f471919c2f6e3a295d0c28040f093e04620f67))
+* **wait:** properly wait for crds introduced mid-deploy ([#5077](https://github.com/zarf-dev/zarf/issues/5077)) ([d15007f](https://github.com/zarf-dev/zarf/commit/d15007fe967ba21c713fb2ac99fbd24f4c7d606a))
+
+## [0.81.0](https://github.com/zarf-dev/zarf/compare/v0.80.0...v0.81.0) (2026-07-09)
+
+
+### ⚠ BREAKING CHANGES
+
+* **ocischeme:** support per-host transport negotiation for plain http ([#5047](https://github.com/zarf-dev/zarf/issues/5047))
+* **sign:** remove legacy signature from production ([#5030](https://github.com/zarf-dev/zarf/issues/5030))
+* **sbom:** remove compare tool ([#5033](https://github.com/zarf-dev/zarf/issues/5033))
+* **verification:** behavior updates for verification logic ([#5011](https://github.com/zarf-dev/zarf/issues/5011))
+
+### Features
+
+* add registry logout function ([#4976](https://github.com/zarf-dev/zarf/issues/4976)) ([c2ab144](https://github.com/zarf-dev/zarf/commit/c2ab1449dcbeeb30598ae73555a11ffd6bc7f6c1))
+* **cmd:** add viper option for cache ([#5051](https://github.com/zarf-dev/zarf/issues/5051)) ([36fd721](https://github.com/zarf-dev/zarf/commit/36fd7213d0cc42721472402d50d2ef1655b018db))
+* deprecate artifact server ([#5041](https://github.com/zarf-dev/zarf/issues/5041)) ([642ac33](https://github.com/zarf-dev/zarf/commit/642ac332ac8d99fde8d4453d197d5a2bf76316df))
+* deprecate sbom viewer ([#5046](https://github.com/zarf-dev/zarf/issues/5046)) ([32e1479](https://github.com/zarf-dev/zarf/commit/32e147910be546a08af7dbf30a41bc003ae4a8f7))
+* **deprecate:** zarf tools registy export ([#5031](https://github.com/zarf-dev/zarf/issues/5031)) ([926c008](https://github.com/zarf-dev/zarf/commit/926c0080c3aacdebac0853c9b343a9c5dd43cf94))
+* direct docker daemon pull ([#5001](https://github.com/zarf-dev/zarf/issues/5001)) ([97c2388](https://github.com/zarf-dev/zarf/commit/97c238802c403ab976c9ed33406797a03c108ee8))
+* group verify flags in CLI help text ([#5008](https://github.com/zarf-dev/zarf/issues/5008)) ([20c32e8](https://github.com/zarf-dev/zarf/commit/20c32e82f3c18d757e8efb0d016e5ab0dbd36d51))
+* **lint:** warn on images without a domain and archive images without .internal ([#4983](https://github.com/zarf-dev/zarf/issues/4983)) ([ed73631](https://github.com/zarf-dev/zarf/commit/ed7363194313dbeeef18cc086c031b0b13be1815))
+* **ocischeme:** support per-host transport negotiation for plain http ([#5047](https://github.com/zarf-dev/zarf/issues/5047)) ([7fc1dec](https://github.com/zarf-dev/zarf/commit/7fc1decb22dc6d555f1d49beefe1e37257088f05))
+* reduce docker daemon pull storage footprint ([#5017](https://github.com/zarf-dev/zarf/issues/5017)) ([f5a443d](https://github.com/zarf-dev/zarf/commit/f5a443d2b6ba70b27d1ece621f0422b0dbec69f7))
+* **sbom:** remove compare tool ([#5033](https://github.com/zarf-dev/zarf/issues/5033)) ([8e4ee75](https://github.com/zarf-dev/zarf/commit/8e4ee75890b4859620c0082b1f755fa5c72b8a80))
+* **schema:** update v1beta1 types ([#4704](https://github.com/zarf-dev/zarf/issues/4704)) ([edc9eb1](https://github.com/zarf-dev/zarf/commit/edc9eb1c4ad8f1e99cbe7397a90c94a7dad25d00))
+* **sign:** remove legacy signature from production ([#5030](https://github.com/zarf-dev/zarf/issues/5030)) ([fab4f62](https://github.com/zarf-dev/zarf/commit/fab4f6265398b07449f7c7a79a8d6576a74fa0d0))
+* **v1beta1:** split target into create time selector and deploy time target ([#5042](https://github.com/zarf-dev/zarf/issues/5042)) ([d188f07](https://github.com/zarf-dev/zarf/commit/d188f07ac6fdf0599f953e63653f3290e3a62002))
+* **verification:** behavior updates for verification logic ([#5011](https://github.com/zarf-dev/zarf/issues/5011)) ([5f3a0a9](https://github.com/zarf-dev/zarf/commit/5f3a0a970ac7102bfd3bb0f4f291de92f8aad5e6))
+
+
+### Bug Fixes
+
+* **flags:** depecate --zarf-cache in favor of just --cache ([#5032](https://github.com/zarf-dev/zarf/issues/5032)) ([9da35b6](https://github.com/zarf-dev/zarf/commit/9da35b677a49af9f346085e51bba3d6cfedc393c))
+* update oras-go and mitigate auto-unpack ([#5049](https://github.com/zarf-dev/zarf/issues/5049)) ([bc9818d](https://github.com/zarf-dev/zarf/commit/bc9818d431022224b52da2a51f7b412ecc2bc3a1))
+
 ## [0.80.0](https://github.com/zarf-dev/zarf/compare/v0.79.0...v0.80.0) (2026-06-25)